wan-service/requirements.txt
fastapi
uvicorn
torch
diffsynth-studio
transformers
accelerate
pillow
python-multipart
